+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/UTxO.hs
@@ -65,7 +65,6 @@ import Cardano.Ledger.Plutus.Data (Data, Datum (..))
 import Cardano.Ledger.Shelley.UTxO (
   getShelleyMinFeeTxUtxo,
   getShelleyWitsVKeyNeeded,
-  shelleyConsumed,
  )
 import Cardano.Ledger.State (
   EraCertState (..),
@@ -99,8 +98,6 @@ deriving instance AlonzoEraScript era => NFData (AlonzoScriptsNeeded era)
 
 instance EraUTxO AlonzoEra where
   type ScriptsNeeded AlonzoEra = AlonzoScriptsNeeded AlonzoEra
-
-  consumed = shelleyConsumed
 
   getConsumedValue = getConsumedMaryValue